chore(web): make setupTests side-effect-only, route fixtures through @/mocks
setupTests.ts now exports nothing — it only wires test lifecycle. The MSW
`server`/`API_BASE` surface moves to a new `@/mocks` barrel that tests
import directly, and the matchMedia machinery moves to
`tests/test-utils/matchMedia.ts` (exposing `installMatchMediaMock` for the
setup-file side effect and `setMobileViewport` for tests).

Repoints the 16 importers from `@/setupTests` to `@/mocks`, surfaces
`setMobileViewport` via `@/tests/test-utils`, and updates the web CLAUDE.md
references.
