+++ web/src/mocks/index.ts
@@ -1,0 +1,2 @@
+export { API_BASE } from './handlers';
+export { server } from './server';
