+++ web/src/tests/integration/league-loader.integration.test.tsx
@@ -1,8 +1,8 @@
 import { League } from '@/components/League/League';
 import type { RouterContext } from '@/lib/router-context';
+import { API_BASE, server } from '@/mocks';
 import { getLeagueById } from '@/services/leagueService';
 import { getLeagueStandings } from '@/services/standingsService';
-import { API_BASE, server } from '@/setupTests';
 import {
   buildAuthenticatedLayout,
   buildTeamRequiredLayout,
