+++ web/src/tests/integration/team-lineup.integration.test.tsx
@@ -1,12 +1,12 @@
 import { MyTeamRoute } from '@/components/Team/Team';
 import type { Team } from '@/contracts/Team';
 import type { RouterContext } from '@/lib/router-context';
+import { API_BASE, server } from '@/mocks';
 import { constructorsQuery } from '@/services/constructorService';
 import { driversQuery } from '@/services/driverService';
 import { getRaceWeekends } from '@/services/raceWeekendService';
 import { seasonQuery } from '@/services/seasonService';
 import { myTeamQuery } from '@/services/teamService';
-import { API_BASE, server } from '@/setupTests';
 import {
   buildAuthenticatedLayout,
   buildStubRoute,
